feat: Per-service tests, Postgres backend, api-version fixes
Tests:
- Split monolithic integration_test.go into 16 per-service files
- All 37 tests pass: go test ./tests/ -run TestBlob, etc.

Postgres backend (DATABASE_URL):
- Backend interface abstraction (Backend in store/backend.go)
- PostgreSQL implementation with JSONB storage (store/postgres.go)
- Auto-creates table on connect, falls back to in-memory if unavailable
- docker-compose.postgres.yml for full-stack local dev

Docs:
- Fixed all ARM curl examples to include ?api-version=2020-06-01
- Removed em dashes from all files
