+++ go.mod
@@ -2,6 +2,8 @@ module github.com/moabukar/miniblue
 
 go 1.23
 
-require github.com/go-chi/chi/v5 v5.1.0
-
-require github.com/google/uuid v1.6.0 // indirect
+require (
+	github.com/go-chi/chi/v5 v5.1.0
+	github.com/google/uuid v1.6.0
+	github.com/lib/pq v1.12.3
+)
