+++ tests/helpers_test.go
@@ -1,0 +1,62 @@
+package tests
+
+import (
+	"bytes"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/moabukar/miniblue/internal/server"
+)
+
+func setupServer() *httptest.Server {
+	srv := server.New()
+	return httptest.NewServer(srv.Handler())
+}
+
+// ---------- helpers ----------
+
+type azError struct {
+	Error struct {
+		Code    string `json:"code"`
+		Message string `json:"message"`
+	} `json:"error"`
+}
+
+func doRequest(t *testing.T, method, url string, body string) *http.Response {
+	t.Helper()
+	var req *http.Request
+	if body != "" {
+		req, _ = http.NewRequest(method, url, bytes.NewBufferString(body))
+		req.Header.Set("Content-Type", "application/json")
+	} else {
+		req, _ = http.NewRequest(method, url, nil)
+	}
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return resp
+}
+
+func expectStatus(t *testing.T, resp *http.Response, want int) {
+	t.Helper()
+	if resp.StatusCode != want {
+		t.Fatalf("expected status %d, got %d", want, resp.StatusCode)
+	}
+}
+
+func decodeJSON(t *testing.T, resp *http.Response) map[string]interface{} {
+	t.Helper()
+	var m map[string]interface{}
+	json.NewDecoder(resp.Body).Decode(&m)
+	return m
+}
+
+func decodeError(t *testing.T, resp *http.Response) azError {
+	t.Helper()
+	var e azError
+	json.NewDecoder(resp.Body).Decode(&e)
+	return e
+}
